+++ packages/nimble-components/karma.conf.js
@@ -25,6 +25,13 @@ const commonChromeFlags = [
     '--force-prefers-reduced-motion'
 ];
 
+// Create a webpack environment plugin to use while running tests so that
+// functionality that accesses the environment, such as the TanStack table
+// within the nimble-table, work correctly.
+// Note: Unless we run the tests twice, we have to choose to either run them
+// against the 'production' configuration or the 'development' configuration.
+// Because we expect shipping apps to use the 'production' configuration, we
+// have chosen to run tests aginst that configuration.
 const webpackEnvironmentPlugin = new webpack.EnvironmentPlugin({
     NODE_ENV: 'production'
 });