+++ packages/nimble-components/src/anchor-button/styles.ts
@@ -1,0 +1,14 @@
+import { css } from '@microsoft/fast-element';
+import {
+    buttonAppearanceVariantStyles,
+    styles as buttonStyles
+} from '../patterns/button/styles';
+
+export const styles = css`
+    ${buttonStyles}
+    ${buttonAppearanceVariantStyles}
+
+    .control {
+        text-decoration: none;
+    }
+`;