+++ packages/nimble-components/src/table/index.ts
@@ -1,17 +1,17 @@
 import { observable } from '@microsoft/fast-element';
 import { DesignSystem, FoundationElement } from '@microsoft/fast-foundation';
 import {
-    ColumnDef,
-    TableState,
-    Updater,
-    Table as TanstackTable,
-    createTable,
-    getCoreRowModel,
-    TableOptionsResolved
+    ColumnDef as TanStackColumnDef,
+    TableState as TanStackTableState,
+    Updater as TanStackUpdater,
+    Table as TanStackTable,
+    createTable as tanStackCreateTable,
+    getCoreRowModel as tanStackGetCoreRowModel,
+    TableOptionsResolved as TanStackTableOptionsResolved
 } from '@tanstack/table-core';
 import { styles } from './styles';
 import { template } from './template';
-import type { TableData } from './types';
+import type { TableRecord } from './types';
 
 declare global {
     interface HTMLElementTagNameMap {
@@ -23,39 +23,36 @@ declare global {
  * A nimble-styled table.
  */
 export class Table<
-    TData extends TableData = {
-        [key: string]: string | number | boolean | Date | null | undefined
-    }
+    TData extends TableRecord = TableRecord
 > extends FoundationElement {
     @observable
     public data: TData[] = [];
 
-    // TODO: Temporarily assume that the data in the table can be represented as strings.
+    // TODO: Temporarily expose the columns as a string array. This will ultimately be
+    // column definitions provided by slotted elements.
     @observable
-    public tableData: string[][] = [];
+    public columns: string[] = [];
 
+    // TODO: Temporarily expose the column headers as a string array.
     @observable
     public columnHeaders: string[] = [];
 
-    private readonly table: TanstackTable<TData>;
-    private options: TableOptionsResolved<TData>;
+    private readonly table: TanStackTable<TData>;
+    private options: TanStackTableOptionsResolved<TData>;
     private readonly tableInitialized: boolean = false;
 
     public constructor() {
         super();
-        const initialData = this.data;
         this.options = {
-            get data(): TData[] {
-                return initialData;
-            },
-            onStateChange: (_: Updater<TableState>) => {},
-            getCoreRowModel: getCoreRowModel(),
+            data: [],
+            onStateChange: (_: TanStackUpdater<TanStackTableState>) => {},
+            getCoreRowModel: tanStackGetCoreRowModel(),
             columns: [],
             state: {},
             renderFallbackValue: null,
             autoResetAll: false
         };
-        this.table = createTable(this.options);
+        this.table = tanStackCreateTable(this.options);
         this.tableInitialized = true;
     }
 
@@ -70,41 +67,25 @@ export class Table<
         // Ignore any updates that occur prior to the TanStack table being initialized.
         if (this.tableInitialized) {
             this.updateTableOptions({ data: this.data });
-            this.refreshRows();
         }
-    }
-
-    // TODO: For now, assume all cells can be rendered as strings. Ultimately, the data
-    // should be passed to nimble-row elements to use the view template from a column definition.
-    private refreshRows(): void {
-        const tableData: string[][] = [];
-        const rows = this.table.getRowModel().rows;
-        for (const row of rows) {
-            const rowArray: string[] = [];
-            for (const cell of row.getVisibleCells()) {
-                rowArray.push(cell.getValue() as string);
-            }
-            tableData.push(rowArray);
-        }
-        this.tableData = tableData;
     }
 
     private updateTableOptions(
-        updatedOptions: Partial<TableOptionsResolved<TData>>
+        updatedOptions: Partial<TanStackTableOptionsResolved<TData>>
     ): void {
         this.options = { ...this.options, ...updatedOptions };
         this.update(this.table.initialState);
     }
 
-    private readonly update = (state: TableState): void => {
+    private readonly update = (state: TanStackTableState): void => {
         this.table.setOptions(prev => ({
             ...prev,
             ...this.options,
             state,
             onStateChange: (updater: unknown) => {
                 const updatedState = typeof updater === 'function'
-                    ? (updater(state) as TableState)
-                    : (updater as TableState);
+                    ? (updater(state) as TanStackTableState)
+                    : (updater as TanStackTableState);
                 this.update(updatedState);
             }
         }));
@@ -120,7 +101,7 @@ export class Table<
         const firstItem = this.data[0]!;
         const keys = Object.keys(firstItem);
         const generatedColumns = keys.map(key => {
-            const columnDef: ColumnDef<TData> = {
+            const columnDef: TanStackColumnDef<TData> = {
                 id: key,
                 accessorKey: key,
                 header: key
@@ -130,6 +111,7 @@ export class Table<
 
         this.updateTableOptions({ columns: generatedColumns });
         this.columnHeaders = generatedColumns.map(x => x.header as string);
+        this.columns = this.columnHeaders;
     }
 }