+++ packages/nimble-components/src/table/tests/table.pageobject.ts
@@ -1,48 +1,58 @@
 import type { Table } from '..';
-import type { TableData } from '../types';
+import type { TableRecord } from '../types';
 
 /**
  * Page object for the `nimble-table` component to provide consistent ways
  * of querying and interacting with the component during tests.
  */
-export class TablePageObject<T extends TableData> {
+export class TablePageObject<T extends TableRecord> {
     public constructor(private readonly tableElement: Table<T>) {}
 
-    public getHeaderContent(columnIndex: number): string {
-        const header = this.tableElement.shadowRoot!.querySelector('.table-header')!;
-        const cells = header.querySelectorAll('.table-cell');
-        if (columnIndex >= cells.length) {
+    public getRenderedHeaderCount(): number {
+        const headers = this.tableElement.shadowRoot!.querySelectorAll(
+            'nimble-table-header'
+        )!;
+        return headers.length;
+    }
+
+    public getRenderedHeaderContent(columnIndex: number): string {
+        const headers = this.tableElement.shadowRoot!.querySelectorAll(
+            'nimble-table-header'
+        )!;
+        if (columnIndex >= headers.length) {
             throw new Error(
-                'Attempting to index passed the totla number of rendered columns'
+                'Attempting to index past the total number of rendered columns'
             );
         }
 
-        return cells.item(columnIndex).textContent || '';
+        return headers.item(columnIndex).textContent?.trim() ?? '';
     }
 
     public getRenderedRowCount(): number {
-        return (
-            this.tableElement.shadowRoot!.querySelectorAll('.table-row')
-                .length || 0
-        );
+        return this.tableElement.shadowRoot!.querySelectorAll(
+            'nimble-table-row'
+        ).length;
     }
 
-    public getCellContent(rowIndex: number, columnIndex: number): string {
-        const rows = this.tableElement.shadowRoot!.querySelectorAll('.table-row');
+    public getRenderedCellContent(
+        rowIndex: number,
+        columnIndex: number
+    ): string {
+        const rows = this.tableElement.shadowRoot!.querySelectorAll('nimble-table-row');
         if (rowIndex >= rows.length) {
             throw new Error(
-                'Attempting to index passed the total number of rendered rows'
+                'Attempting to index past the total number of rendered rows'
             );
         }
 
         const row = rows.item(rowIndex);
-        const cells = row.querySelectorAll('.table-cell');
+        const cells = row.shadowRoot!.querySelectorAll('nimble-table-cell');
         if (columnIndex >= cells.length) {
             throw new Error(
-                'Attempting to index passed the totla number of rendered columns'
+                'Attempting to index past the total number of rendered columns'
             );
         }
 
-        return cells.item(columnIndex).textContent || '';
+        return cells.item(columnIndex).shadowRoot!.textContent?.trim() ?? '';
     }
 }