+++ packages/nimble-components/src/table/tests/table.spec.ts
@@ -2,59 +2,69 @@ import { html } from '@microsoft/fast-element';
 import { Table } from '..';
 import { waitForUpdatesAsync } from '../../testing/async-helpers';
 import { type Fixture, fixture } from '../../utilities/tests/fixture';
-import type { TableData } from '../types';
+import type { TableRecord } from '../types';
 import { TablePageObject } from './table.pageobject';
 
-interface FakeTableData extends TableData {
+interface SimpleTableRecord extends TableRecord {
     stringData: string;
     numericData: number;
     booleanData: boolean;
     dateData: Date;
 }
-const fakeTableDataKeys = [
+
+const simpleTableDataKeys = [
     'stringData',
     'numericData',
     'booleanData',
     'dateData'
-];
+] as const;
 
-async function setup(): Promise<Fixture<Table<FakeTableData>>> {
-    return fixture<Table<FakeTableData>>(html`<nimble-table></nimble-table>`);
-}
-
-function makeFakeData(count: number): FakeTableData[] {
-    const data: FakeTableData[] = [];
-    for (let i = 0; i < count; i++) {
-        data.push({
-            stringData: `string #${i}`,
-            numericData: i * 2,
-            booleanData: i % 2 === 0,
-            // Generate arbitrary dates that are 10 seconds apart
-            dateData: new Date(1670000000000 + i * 10000)
-        });
+const simpleTableData = [
+    {
+        stringData: 'string 1',
+        numericData: 8,
+        booleanData: true,
+        dateData: new Date(2008, 12, 11)
+    },
+    {
+        stringData: 'hello world',
+        numericData: 0,
+        booleanData: true,
+        dateData: new Date(2022, 5, 30)
+    },
+    {
+        stringData: 'string 1',
+        numericData: -9,
+        booleanData: false,
+        dateData: new Date(2000, 1, 1)
     }
-    return data;
+] as const;
+
+async function setup(): Promise<Fixture<Table<SimpleTableRecord>>> {
+    return fixture<Table<SimpleTableRecord>>(
+        html`<nimble-table></nimble-table>`
+    );
 }
 
 describe('Table', () => {
-    let element: Table<FakeTableData>;
+    let element: Table<SimpleTableRecord>;
     let connect: () => Promise<void>;
     let disconnect: () => Promise<void>;
-    let pageObject: TablePageObject<FakeTableData>;
+    let pageObject: TablePageObject<SimpleTableRecord>;
 
-    function verifyRenderedData(expectedData: FakeTableData[]): void {
+    function verifyRenderedData(expectedData: SimpleTableRecord[]): void {
         const expectedRowCount = expectedData.length;
         expect(pageObject.getRenderedRowCount()).toEqual(expectedRowCount);
         for (let rowIndex = 0; rowIndex < expectedRowCount; rowIndex++) {
             for (
                 let columnIndex = 0;
-                columnIndex < fakeTableDataKeys.length;
+                columnIndex < simpleTableDataKeys.length;
                 columnIndex++
             ) {
-                const dataKey = fakeTableDataKeys[columnIndex]!;
+                const dataKey = simpleTableDataKeys[columnIndex]!;
                 const expectedCellData = expectedData[rowIndex]![dataKey]!;
                 expect(
-                    pageObject.getCellContent(rowIndex, columnIndex)
+                    pageObject.getRenderedCellContent(rowIndex, columnIndex)
                 ).toEqual(expectedCellData.toString());
             }
         }
@@ -62,7 +72,7 @@ describe('Table', () => {
 
     beforeEach(async () => {
         ({ element, connect, disconnect } = await setup());
-        pageObject = new TablePageObject<FakeTableData>(element);
+        pageObject = new TablePageObject<SimpleTableRecord>(element);
     });
 
     afterEach(async () => {
@@ -76,48 +86,49 @@ describe('Table', () => {
     it('should render column headers', async () => {
         await connect();
 
-        const fakeData = makeFakeData(1);
-        element.data = fakeData;
+        element.data = [...simpleTableData];
         await waitForUpdatesAsync();
 
+        expect(pageObject.getRenderedHeaderCount()).toEqual(
+            simpleTableDataKeys.length
+        );
         for (
             let columnIndex = 0;
-            columnIndex < fakeTableDataKeys.length;
+            columnIndex < simpleTableDataKeys.length;
             columnIndex++
         ) {
-            expect(pageObject.getHeaderContent(columnIndex)).toEqual(
-                fakeTableDataKeys[columnIndex]!
+            expect(pageObject.getRenderedHeaderContent(columnIndex)).toEqual(
+                simpleTableDataKeys[columnIndex]!
             );
         }
     });
 
     it('can set data before the element is connected', async () => {
-        const fakeData = makeFakeData(5);
-        element.data = fakeData;
+        const data = [...simpleTableData];
+        element.data = data;
         await connect();
 
-        verifyRenderedData(fakeData);
+        verifyRenderedData(data);
     });
 
     it('can set data after the element is connected', async () => {
         await connect();
 
-        const fakeData = makeFakeData(5);
-        element.data = fakeData;
+        const data = [...simpleTableData];
+        element.data = data;
         await waitForUpdatesAsync();
 
-        verifyRenderedData(fakeData);
+        verifyRenderedData(data);
     });
 
     it('updating data can add a new row to the table', async () => {
         await connect();
 
-        const fakeData = makeFakeData(5);
-        element.data = fakeData;
+        element.data = [...simpleTableData];
         await waitForUpdatesAsync();
 
-        const updatedData: FakeTableData[] = [
-            ...fakeData,
+        const updatedData: SimpleTableRecord[] = [
+            ...simpleTableData,
             {
                 stringData: 'a new string',
                 numericData: -9,
@@ -134,14 +145,12 @@ describe('Table', () => {
     it('updating data can remove rows from the table', async () => {
         await connect();
 
-        const fakeData = makeFakeData(5);
-        element.data = fakeData;
+        element.data = [...simpleTableData];
         await waitForUpdatesAsync();
 
-        const updatedData: FakeTableData[] = [
-            fakeData[0]!,
-            fakeData[1]!,
-            fakeData[4]!
+        const updatedData: SimpleTableRecord[] = [
+            simpleTableData[0],
+            simpleTableData[2]
         ];
         element.data = updatedData;
         await waitForUpdatesAsync();
@@ -152,16 +161,13 @@ describe('Table', () => {
     it('updating data can reorder rows from the table', async () => {
         await connect();
 
-        const fakeData = makeFakeData(5);
-        element.data = fakeData;
+        element.data = [...simpleTableData];
         await waitForUpdatesAsync();
 
-        const updatedData: FakeTableData[] = [
-            fakeData[4]!,
-            fakeData[1]!,
-            fakeData[3]!,
-            fakeData[2]!,
-            fakeData[0]!
+        const updatedData: SimpleTableRecord[] = [
+            simpleTableData[1],
+            simpleTableData[2],
+            simpleTableData[0]
         ];
         element.data = updatedData;
         await waitForUpdatesAsync();