+++ packages/nimble-components/src/theme-provider/design-tokens.ts
@@ -84,7 +84,12 @@ import {
     ButtonLabel1LineHeight,
     TooltipCaptionLineHeight,
     Information100LightUi,
-    Information100DarkUi
+    Information100DarkUi,
+    DigitalGreenDark,
+    PowerGreen,
+    GridHeaderFamily,
+    GridHeaderWeight,
+    GridHeaderSize
 } from '@ni/nimble-tokens/dist/styledictionary/js/tokens';
 import {
     modalBackdropColorThemeColorStatic,
@@ -185,6 +190,10 @@ export const popupBorderColor = DesignToken.create<string>(
 export const tooltipBackgroundColor = DesignToken.create<string>(
     styleNameFromTokenName(tokenNames.tooltipBackgroundColor)
 ).withDefault((element: HTMLElement) => getColorForTheme(element, Black15, Black85, ForestGreen));
+
+export const tableRowBorderColor = DesignToken.create<string>(
+    styleNameFromTokenName(tokenNames.tableRowBorderColor)
+).withDefault((element: HTMLElement) => getColorForTheme(element, Black15, Black88, ForestGreen));
 
 // Component Sizing Tokens
 export const controlHeight = DesignToken.create<string>(
@@ -351,16 +360,81 @@ export const [
 );
 
 export const [
-    linkStandardFont,
-    linkStandardFontColor,
-    linkStandardDisabledFontColor,
-    linkStandardFontFamily,
-    linkStandardFontWeight,
-    linkStandardFontSize,
-    linkStandardFontLineHeight,
-    linkStandardFallbackFontFamily
+    linkFont,
+    linkFontColor,
+    linkDisabledFontColor,
+    linkFontFamily,
+    linkFontWeight,
+    linkFontSize,
+    linkFontLineHeight,
+    linkFallbackFontFamily
 ] = createFontTokens(
-    tokenNames.linkStandardFont,
+    tokenNames.linkFont,
+    (element: HTMLElement) => getDefaultFontColorForTheme(element),
+    (element: HTMLElement) => hexToRgbaCssColor(getDefaultFontColorForTheme(element), 0.3),
+    LinkLightUiFamily,
+    LinkLightUiWeight,
+    LinkLightUiSize,
+    LinkLineHeight,
+    'sans-serif'
+);
+
+export const [
+    linkActiveFont,
+    linkActiveFontColor,
+    linkActiveDisabledFontColor,
+    linkActiveFontFamily,
+    linkActiveFontWeight,
+    linkActiveFontSize,
+    linkActiveFontLineHeight,
+    linkActiveFallbackFontFamily
+] = createFontTokens(
+    tokenNames.linkActiveFont,
+    (element: HTMLElement) => getColorForTheme(
+        element,
+        DigitalGreenDark,
+        PowerGreen,
+        hexToRgbaCssColor(White, 0.6)
+    ),
+    (element: HTMLElement) => hexToRgbaCssColor(getDefaultFontColorForTheme(element), 0.3),
+    LinkLightUiFamily,
+    LinkLightUiWeight,
+    LinkLightUiSize,
+    LinkLineHeight,
+    'sans-serif'
+);
+
+export const [
+    linkProminentFont,
+    linkProminentFontColor,
+    linkProminentDisabledFontColor,
+    linkProminentFontFamily,
+    linkProminentFontWeight,
+    linkProminentFontSize,
+    linkProminentFontLineHeight,
+    linkProminentFallbackFontFamily
+] = createFontTokens(
+    tokenNames.linkProminentFont,
+    (element: HTMLElement) => getColorForTheme(element, DigitalGreenDark, PowerGreen, PowerGreen),
+    (element: HTMLElement) => hexToRgbaCssColor(getDefaultFontColorForTheme(element), 0.3),
+    LinkLightUiFamily,
+    LinkLightUiWeight,
+    LinkLightUiSize,
+    LinkLineHeight,
+    'sans-serif'
+);
+
+export const [
+    linkActiveProminentFont,
+    linkActiveProminentFontColor,
+    linkActiveProminentDisabledFontColor,
+    linkActiveProminentFontFamily,
+    linkActiveProminentFontWeight,
+    linkActiveProminentFontSize,
+    linkActiveProminentFontLineHeight,
+    linkActiveProminentFallbackFontFamily
+] = createFontTokens(
+    tokenNames.linkActiveProminentFont,
     (element: HTMLElement) => getDefaultFontColorForTheme(element),
     (element: HTMLElement) => hexToRgbaCssColor(getDefaultFontColorForTheme(element), 0.3),
     LinkLightUiFamily,
@@ -526,6 +600,26 @@ export const [
     ErrorLightUiFamily,
     ErrorLightUiWeight,
     ErrorLightUiSize,
+    TooltipCaptionLineHeight,
+    'sans-serif'
+);
+
+export const [
+    tableHeaderFont,
+    tableHeaderFontColor,
+    tableHeaderDisabledFontColor,
+    tableHeaderFontFamily,
+    tableHeaderFontWeight,
+    tableHeaderFontSize,
+    tableHeaderFontLineHeight,
+    tableHeaderFallbackFontFamily
+] = createFontTokens(
+    tokenNames.tableHeaderFont,
+    (element: HTMLElement) => hexToRgbaCssColor(getDefaultFontColorForTheme(element), 0.6),
+    (element: HTMLElement) => hexToRgbaCssColor(getDefaultFontColorForTheme(element), 0.3),
+    GridHeaderFamily,
+    GridHeaderWeight,
+    GridHeaderSize,
     TooltipCaptionLineHeight,
     'sans-serif'
 );